refactor: moving process_code_and_sourcemap to utils
